requirements.txt
streamlit
llama-index
llama-index-llms-cohere
llama-index-embeddings-cohere
cohere
python-dotenv
pandas
openpyxl
# Additional dependencies for the project